feat: optimized comics have a zip comment in them to prevent any future optimize

+++ lib/basenji/reader/process/comic_optimizer.ex
@@ -1,44 +1,61 @@
 defmodule Basenji.Reader.Process.ComicOptimizer do
   @moduledoc false
 
-  import Basenji.Reader
+  alias Basenji.Reader
 
-  def optimize(comic_file_path) do
-    tmp_dir =
-      System.tmp_dir!()
-      |> Path.join("basenji")
-      |> Path.join("basenji_optimize_#{System.monotonic_time()}")
+  def optimize(comic_file_path, result_directory) do
+    if String.ends_with?(comic_file_path, "optimized.cbz") || basenji_comment?(comic_file_path) do
+      {:ok, comic_file_path}
+    else
+      :ok = File.mkdir_p!(result_directory)
 
-    :ok = File.mkdir_p!(tmp_dir)
+      comic_name =
+        Path.basename(comic_file_path)
+        |> Path.rootname()
+        |> ProperCase.snake_case()
 
-    comic_name =
-      Path.basename(comic_file_path)
-      |> Path.rootname()
-      |> ProperCase.snake_case()
+      images_dir = Path.join(result_directory, "#{comic_name}_images")
+      :ok = File.mkdir_p!(images_dir)
 
-    images_dir = Path.join(tmp_dir, "#{comic_name}_images")
-    :ok = File.mkdir_p!(images_dir)
+      with {:ok, %{page_count: page_count}} <- Reader.info(comic_file_path),
+           {:ok, %{entries: entries}} <- Reader.read(comic_file_path),
+           {:ok, stream} <- Reader.stream_pages(comic_file_path, optimize: true) do
+        padding = String.length("#{page_count}") - 1
 
-    with {:ok, %{page_count: page_count}} <- info(comic_file_path),
-         {:ok, %{entries: entries}} <- read(comic_file_path),
-         {:ok, stream} <- stream_pages(comic_file_path, optimize: true) do
-      padding = String.length("#{page_count}") - 1
+        stream
+        |> Stream.with_index()
+        |> Enum.each(fn {page, page_idx} ->
+          page_bytes = page |> Enum.to_list()
+          ext = Enum.at(entries, page_idx) |> Map.get(:file_name) |> Path.extname()
 
-      stream
-      |> Stream.with_index()
-      |> Enum.each(fn {page, page_idx} ->
-        page_bytes = page |> Enum.to_list()
-        ext = Enum.at(entries, page_idx) |> Map.get(:file_name) |> Path.extname()
+          file_name = "#{String.pad_leading("#{page_idx + 1}", padding, "0")}#{ext}"
+          :ok = File.write!(Path.join(images_dir, file_name), page_bytes)
+        end)
 
-        file_name = "#{String.pad_leading("#{page_idx + 1}", padding, "0")}#{ext}"
-        :ok = File.write!(Path.join(images_dir, file_name), page_bytes)
-      end)
+        optimized_name = "#{comic_name}_optimized.cbz"
 
-      optimized_name = "#{comic_name}_optimized.cbz"
+        response = zip(result_directory, "#{comic_name}_images", optimized_name)
+        File.rm_rf!(images_dir)
+        response
+      end
+    end
+  end
 
-      response = zip(tmp_dir, "#{comic_name}_images", optimized_name)
-      File.rm_rf!(images_dir)
-      response
+  def basenji_comment?(comic_file_path) do
+    {:ok, %{format: format}} = Reader.info(comic_file_path)
+
+    if format == :cbz do
+      Reader.exec("zipinfo", ["-z", comic_file_path])
+      |> case do
+        {:ok, output} ->
+          String.contains?(output, "Optimized by Basenji")
+
+        other ->
+          IO.inspect(other)
+          false
+      end
+    else
+      false
     end
   end
 
@@ -49,15 +66,14 @@ defmodule Basenji.Reader.Process.ComicOptimizer do
       # Maximum compression level
       "-9",
       # Quiet
-      "-q",
-      # Use deflate compression method
-      "-Z",
-      "bzip2"
+      "-q"
     ]
 
     System.cmd("zip", flags ++ [zip_file_name, images_dir_name], cd: parent_dir)
     |> case do
       {_, 0} ->
+        comment = "Optimized by Basenji v#{Application.spec(:basenji, :vsn)} on #{DateTime.utc_now()}"
+        System.cmd("sh", ["-c", "echo '#{comment}' | zip -z #{zip_file_name}"], cd: parent_dir)
         {:ok, Path.join(parent_dir, zip_file_name)}
 
       {error_output, exit_code} ->
